requirements.txt
wheel
djangorestframework
markdown
django-filter
django
psycopg2
requests
django-background-tasks
pytest
selenium
webdriver_manager
pyyaml
djangorestframework-recursive
boto3
ebooklib
pyyaml
Pillow
mutagen
pydub
elasticsearch
elasticsearch_dsl